Add helmert encoder test file and rename polynomial and backwards difference test file to match test_{fillename} under test
