+++ category_encoders/tests/test_polynomial.py
